+++ ocpp-v1_6/src/main/java/eu/chargetime/ocpp/model/core/StartTransactionConfirmation.java
@@ -29,6 +29,8 @@ package eu.chargetime.ocpp.model.core;
 
 import eu.chargetime.ocpp.model.Confirmation;
 import eu.chargetime.ocpp.utilities.MoreObjects;
+
+import java.math.BigInteger;
 import java.util.Objects;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
@@ -42,10 +44,10 @@ import javax.xml.bind.annotation.XmlType;
 public class StartTransactionConfirmation implements Confirmation {
 
   private IdTagInfo idTagInfo;
-  private Long transactionId;
+  private BigInteger transactionId;
 
   /**
-   * @deprecated use {@link #StartTransactionConfirmation(IdTagInfo, Integer)} to be sure to set
+   * @deprecated use {@link #StartTransactionConfirmation(IdTagInfo, BigInteger)} to be sure to set
    *     required fields
    */
   @Deprecated
@@ -55,9 +57,9 @@ public class StartTransactionConfirmation implements Confirmation {
    * Handle required fields.
    *
    * @param idTagInfo the {@link IdTagInfo}, see {@link #setIdTagInfo(IdTagInfo)}
-   * @param transactionId integer, transaction, see {@link #setTransactionId(Integer)}
+   * @param transactionId integer, transaction, see {@link #setTransactionId(BigInteger)}
    */
-  public StartTransactionConfirmation(IdTagInfo idTagInfo, Long transactionId) {
+  public StartTransactionConfirmation(IdTagInfo idTagInfo, BigInteger transactionId) {
     setIdTagInfo(idTagInfo);
     setTransactionId(transactionId);
   }
@@ -94,7 +96,7 @@ public class StartTransactionConfirmation implements Confirmation {
    *
    * @return transaction id.
    */
-  public Long getTransactionId() {
+  public BigInteger getTransactionId() {
     return transactionId;
   }
 
@@ -104,7 +106,7 @@ public class StartTransactionConfirmation implements Confirmation {
    * @param transactionId integer, transaction.
    */
   @XmlElement
-  public void setTransactionId(Long transactionId) {
+  public void setTransactionId(BigInteger transactionId) {
     this.transactionId = transactionId;
   }
